# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: personal-finance
  version: "1.2.0"
  description: |
    Localhost-only HTTP API for the personal-finance backend.

    **Single source of truth contract (Phase 1.8, 2026-04-20):** this file is
    the canonical contract between `personal-finance` (Go) and `personal-web`
    (TypeScript). Hand-written enums in `internal/handler/validate.go` and
    response DTOs in `internal/handler/dto.go` MUST stay in sync with the
    schemas below. The web client uses `openapi-typescript` to regenerate
    `src/types.gen.ts` from this file; any drift produces a type error in
    the web build.

    Maintenance protocol when changing the API:
      1. Edit Go: validate.go / dto.go / handler.go
      2. Edit this file to match
      3. `cd ~/personal-web && npm run gen:api` regenerates types.gen.ts
      4. `cd ~/personal-web && npm run build` should pass

    Why hand-written instead of generated from Go:
      - Go has no production-grade "AST -> OpenAPI" tool that handles
        Gin's untyped `c.JSON(gin.H{...})` shape.
      - The endpoint count (~17) is small enough that hand-maintenance is
        cheaper than introducing swaggo annotations on every handler.
      - When the count crosses ~25 we should revisit (likely swaggo or
        moving handlers to typed responses + ogen).
servers:
  - url: http://127.0.0.1:7001
    description: local dev (default)
...
paths:
...
  /api/finance/assets:
    get:
      tags: [assets]
      summary: List assets
      operationId: listAssets
      parameters:
        - { in: query, name: bucket,           schema: { $ref: "#/components/schemas/Bucket" } }
        - { in: query, name: asset_type,       schema: { $ref: "#/components/schemas/AssetType" } }
        - { in: query, name: include_archived, schema: { type: string, enum: ["1", "true"] } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [assets, count]
                properties:
                  assets: { type: array, items: { $ref: "#/components/schemas/Asset" } }
                  count:  { type: integer }
        "400": { $ref: "#/components/responses/BadRequest" }
        "500": { $ref: "#/components/responses/InternalError" }
    post:
      tags: [assets]
      summary: Upsert asset by code (idempotent)
      operationId: upsertAsset
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpsertAssetReq" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [asset]
                properties:
                  asset: { $ref: "#/components/schemas/Asset" }
        "400": { $ref: "#/components/responses/BadRequest" }
  /api/finance/assets/{id}:
    parameters:
      - { in: path, name: id, required: true, schema: { type: integer, format: int64 } }
    get:
      tags: [assets]
      summary: Get asset by id
      operationId: getAsset
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [asset]
                properties:
                  asset: { $ref: "#/components/schemas/Asset" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      tags: [assets]
      summary: Patch asset (partial update)
      operationId: patchAsset
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PatchAssetReq" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [asset]
                properties:
                  asset: { $ref: "#/components/schemas/Asset" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [assets]
      summary: Archive asset (soft delete)
      operationId: archiveAsset
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [archived_id]
                properties:
                  archived_id: { type: integer, format: int64 }
        "404": { $ref: "#/components/responses/NotFound" }

  # ---------- snapshots ----------
  /api/finance/snapshots:
    get:
      tags: [snapshots]
      summary: List snapshots
      operationId: listSnapshots
      parameters:
        - { in: query, name: asset_id, schema: { type: integer, format: int64 } }
        - { in: query, name: since,    schema: { type: string, format: date } }
        - { in: query, name: until,    schema: { type: string, format: date } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [snapshots, count]
                properties:
                  snapshots: { type: array, items: { $ref: "#/components/schemas/Snapshot" } }
                  count:     { type: integer }
    post:
      tags: [snapshots]
      summary: Upsert snapshot by (asset_id, snapshot_date)
      operationId: upsertSnapshot
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpsertSnapshotReq" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [snapshot]
                properties:
                  snapshot: { $ref: "#/components/schemas/Snapshot" }
        "400": { $ref: "#/components/responses/BadRequest" }
  /api/finance/snapshots/{id}:
    parameters:
      - { in: path, name: id, required: true, schema: { type: integer, format: int64 } }
    get:
      tags: [snapshots]
      summary: Get snapshot by id
      operationId: getSnapshot
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [snapshot]
                properties:
                  snapshot: { $ref: "#/components/schemas/Snapshot" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      tags: [snapshots]
      summary: Patch snapshot
      operationId: patchSnapshot
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PatchSnapshotReq" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [snapshot]
                properties:
                  snapshot: { $ref: "#/components/schemas/Snapshot" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [snapshots]
      summary: Delete snapshot
      operationId: deleteSnapshot
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [deleted_id]
                properties:
                  deleted_id: { type: integer, format: int64 }
        "404": { $ref: "#/components/responses/NotFound" }

  # ---------- transactions ----------
  /api/finance/transactions:
    get:
      tags: [transactions]
      summary: List transactions
      operationId: listTransactions
      parameters:
        - { in: query, name: asset_id,  schema: { type: integer, format: int64 } }
        - { in: query, name: direction, schema: { $ref: "#/components/schemas/Direction" } }
        - { in: query, name: since,     schema: { type: string, format: date } }
        - { in: query, name: until,     schema: { type: string, format: date } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [transactions, count]
                properties:
                  transactions: { type: array, items: { $ref: "#/components/schemas/Transaction" } }
                  count:        { type: integer }
    post:
      tags: [transactions]
      summary: Create transaction (events are immutable; no UPSERT)
      operationId: createTransaction
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateTransactionReq" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                required: [transaction]
                properties:
                  transaction: { $ref: "#/components/schemas/Transaction" }
  /api/finance/transactions/{id}:
    parameters:
      - { in: path, name: id, required: true, schema: { type: integer, format: int64 } }
    get:
      tags: [transactions]
      summary: Get transaction by id
      operationId: getTransaction
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [transaction]
                properties:
                  transaction: { $ref: "#/components/schemas/Transaction" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      tags: [transactions]
      summary: Patch transaction
      operationId: patchTransaction
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PatchTransactionReq" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [transaction]
                properties:
                  transaction: { $ref: "#/components/schemas/Transaction" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [transactions]
      summary: Delete transaction
      operationId: deleteTransaction
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [deleted_id]
                properties:
                  deleted_id: { type: integer, format: int64 }
        "404": { $ref: "#/components/responses/NotFound" }

  # ---------- holdings ----------
...
  /api/finance/bucket-targets:
    get:
      tags: [bucket-targets]
      summary: List target allocations (always returns 3 buckets)
      operationId: listBucketTargets
      description: |
        Returns one entry per bucket (cash / stable / growth) regardless of
        whether the row has been set yet. `is_set=false` + `target_pct=null`
        means the user hasn't picked a target for that bucket. `sum_pct` is
        the sum across set targets, or `null` if none are set.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [bucket_targets, sum_pct]
                properties:
                  bucket_targets:
                    type: array
                    minItems: 3
                    maxItems: 3
                    items: { $ref: "#/components/schemas/BucketTarget" }
                  sum_pct: { type: number, nullable: true }
    put:
      tags: [bucket-targets]
      summary: Upsert one bucket target (idempotent on bucket)
      operationId: upsertBucketTarget
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpsertBucketTargetReq" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [bucket_target]
                properties:
                  bucket_target: { $ref: "#/components/schemas/BucketTarget" }
        "400": { $ref: "#/components/responses/BadRequest" }
  /api/finance/bucket-targets/{bucket}:
    parameters:
      - { in: path, name: bucket, required: true, schema: { $ref: "#/components/schemas/Bucket" } }
    delete:
      tags: [bucket-targets]
      summary: Clear target for one bucket
      operationId: deleteBucketTarget
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [deleted_bucket]
                properties:
                  deleted_bucket: { $ref: "#/components/schemas/Bucket" }
        "404": { $ref: "#/components/responses/NotFound" }

  # ---------- publish ----------
...
components:
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiError" }
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiError" }
  schemas:
...
    AssetType:
      type: string
      enum:
        - cash-cny
        - wealth-mgmt-product
        - etf-fund
        - cn-stock
        - hk-stock
        - us-stock
        - social-insurance
        - real-estate
    Bucket:
      type: string
      enum: [cash, stable, growth]
...